requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blurhash==1.1.4
boost==0.1
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dlib @ file:///C:/Users/24079/dlib-19.23.0-cp39-cp39-win_amd64.whl
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.88.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.8.1
flatbuffers==22.11.23
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
GCT==1.0
gdown==4.6.0
geographiclib==2.0
geopy==2.3.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.12.0
greenlet==1.1.3.post0
grpcio==1.50.0
h11==0.12.0
h5py==3.7.0
httpcore==0.15.0
httpx==0.23.1
idna==2.10
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.4.4
libclang==14.0.6
linkify-it-py==1.0.3
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
Mastodon.py==1.5.1
mat73==0.60
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
mistune==2.0.4
multidict==6.0.3
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.1
notebook-shim==0.1.0
numpy==1.21.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
orjson==3.8.3
outcome==1.2.0
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pydub==0.25.1
pygame==2.1.2
Pygments==2.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.5
pytorch-model-summary==0.1.2
pytz==2022.2.1
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.25.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
selenium==4.8.2
Send2Trash==1.8.0
shapely==2.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
stack-data==0.5.1
starlette==0.22.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.16.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.10.0
typing_extensions==4.3.0
uc-micro-py==1.0.1
urllib3==1.26.12
uvicorn==0.20.0
visdom==0.2.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
yarl==1.8.2
zipp==3.9.0
